Labs/GCP_Labs/Flight_Delay_Prediction_Pipeline_Lab/cloud-functions/requirements.txt
functions-framework==3.*
google-auth
google-auth-oauthlib
google-auth-httplib2
requests